+++ bt_device.py
@@ -5,6 +5,7 @@
 Automatically checks for new usb drives to perform the update of the SM Demo Software
 
 Changelog:
+- 1.2.3 - make device pairable by default
 - 1.2.2 - added more log outputs
 - 1.2.1 - enhanced auto-bluetooth recovery
 - 1.2.0 - restart bluetooth service if too many attempts are made
@@ -21,7 +22,7 @@ __author__ = "Alberto Occelli"
 __copyright__ = "Copyright 2023,"
 __credits__ = ["Alberto Occelli"]
 __license__ = "MIT"
-__version__ = "1.2.2"
+__version__ = "1.2.3"
 __maintainer__ = "Alberto Occelli"
 __email__ = "albertoccelli@gmail.com"
 __status__ = "Dev"
@@ -32,6 +33,7 @@ import time
 from utils import get_sinks, print_datetime, curwd
 from config import *
 
+lang = "eng"
 
 class Device:
 
@@ -45,14 +47,18 @@ class Device:
         self.ready_to_play = False
         self.get_mac_address()
         self.get_info()
+        pairable = subprocess.Popen(["bluetoothctl", "pairable", "on"])
+        pairable.wait()
         if not self.trusted:
             self.trust()
         if not self.paired:
             self.pair()
         if not self.connected:
             self.connect()
+        '''
         else:
             print_datetime("Device already connected")
+        '''
 
     def get_mac_address(self):
         ntry = 1
@@ -62,6 +68,7 @@ class Device:
         get_devices = subprocess.run(["bluetoothctl", "devices"], capture_output=True, text=True)
         device_list = get_devices.stdout.split("\n")
         device_list.remove("")
+        print_datetime(f"Looking for device named {self.name}")
         for i in range(len(device_list)):
             if self.name in device_list[i]:
                 self.mac_address = device_list[i].split(" ")[1]
@@ -119,9 +126,10 @@ class Device:
                     else:
                         self.connected = False
 
-    def connect(self):
+    def connect(self, mute=False):
         attempts = 0
-        audio_prompt(f"{curwd}/prompts/{lang}/attempt.wav")
+        if not mute:
+            audio_prompt(f"{curwd}/prompts/{lang}/attempt.wav")
         if self.mac_address:
             while True:
                 print_datetime("Trying to connect...")
@@ -168,8 +176,7 @@ class Device:
                     print_datetime("Failed to pair. Please try putting the device into pairing mode")
                     audio_prompt(f"{curwd}/prompts/{lang}/error1.wav")
             elif "success" in outcome.lower():
-                audio_prompt(f"{curwd}/prompts/{lang}/connected.wav")
-                self.get_sink()
+                self.get_sink("pairing")
                 return
             attempts += 1
 
@@ -180,14 +187,14 @@ class Device:
             outcome = trust.stdout
             print_datetime(outcome)
 
-    def get_sink(self):
+    def get_sink(self, message = None):
         sinks = get_sinks()
         try:
             self.sink = sinks[1]
             if len(sinks) == 2:
                 self.ready_to_play = True
         except Exception as e:
-            print_datetime(f"Error getting sinks: {e}")
+            print_datetime(f"Error getting sinks during {message}: {e}")
             self.ready_to_play = False
             if not self.check_connected():
                 self.connect()
@@ -216,4 +223,5 @@ if __name__ == "__main__":
 
     # initialize the neckband connection. Closes the program once it's ready to play
     neckband = Device(device_name)
+    neckband.connect(mute=True)
     initialize(neckband)